\name{enums}
\alias{PolyBlockType-class}
\alias{Orientation-class}
\alias{WritingDirection-class}
\alias{TextlineOrder-class}
\alias{PageSegMode-class}
%\alias{PageIteratorLevel-class}
\alias{ParagraphJustification-class}
\alias{OcrEngineMode-class}
\alias{StrongScriptDirection-class}
\alias{coerce,character,OcrEngineMode-method}
\alias{coerce,character,Orientation-method}
\alias{coerce,character,PageSegMode-method}
\alias{coerce,character,ParagraphJustification-method}
\alias{coerce,character,PolyBlockType-method}
\alias{coerce,character,StrongScriptDirection-method}
\alias{coerce,character,TextlineOrder-method}
\alias{coerce,character,WritingDirection-method}
\alias{coerce,integer,OcrEngineMode-method}
\alias{coerce,integer,Orientation-method}
\alias{coerce,integer,PageIteratorLevel-method}
\alias{coerce,integer,PageSegMode-method}
\alias{coerce,integer,ParagraphJustification-method}
\alias{coerce,integer,PolyBlockType-method}
\alias{coerce,integer,StrongScriptDirection-method}
\alias{coerce,integer,TextlineOrder-method}
\alias{coerce,integer,WritingDirection-method}
\alias{coerce,numeric,OcrEngineMode-method}
\alias{coerce,numeric,Orientation-method}
\alias{coerce,numeric,PageSegMode-method}
\alias{coerce,numeric,ParagraphJustification-method}
\alias{coerce,numeric,PolyBlockType-method}
\alias{coerce,numeric,StrongScriptDirection-method}
\alias{coerce,numeric,TextlineOrder-method}
\alias{coerce,numeric,WritingDirection-method}


\alias{PSM_OSD_ONLY}
\alias{PSM_AUTO_OSD}
\alias{PSM_AUTO_ONLY}
\alias{PSM_AUTO}
\alias{PSM_SINGLE_COLUMN}
\alias{PSM_SINGLE_BLOCK_VERT_TEXT}
\alias{PSM_SINGLE_BLOCK}
\alias{PSM_SINGLE_LINE}
\alias{PSM_SINGLE_WORD}
\alias{PSM_CIRCLE_WORD}
\alias{PSM_SINGLE_CHAR}
\alias{PSM_SPARSE_TEXT}
\alias{PSM_SPARSE_TEXT_OSD}
\alias{PSM_RAW_LINE}
\alias{PSM_COUNT}

\alias{RIL_BLOCK}
\alias{RIL_PARA}
\alias{RIL_TEXTLINE}
\alias{RIL_WORD}
\alias{RIL_SYMBOL}


\alias{OEM_TESSERACT_ONLY}
\alias{OEM_CUBE_ONLY}
\alias{OEM_TESSERACT_CUBE_COMBINED}
\alias{OEM_DEFAULT}

\alias{OEM_LSTM_ONLY}
\alias{OEM_TESSERACT_LSTM_COMBINED}

\alias{InputFileFormat-class}
\alias{coerce,character,InputFileFormat-method}
\alias{coerce,integer,InputFileFormat-method}
\alias{coerce,numeric,InputFileFormat-method}

\title{Tesseract Enums}
\description{
Enumerated constant classes used in the Tesseract API.
}
