\name{opaqueClasses-class}
\Rdversion{1.1}
\docType{class}
\alias{ResultIterator-class}
\alias{TessPDFRenderer-class}
\title{Opaque Classes for Rtesseract}
\description{
  These classes are not intended for direct access by an R user/programmer.
  Instead, they are exported by Rtesseract so the C code can create
  instances of them and they are used internally within unexported R
  code in the package
}
\keyword{classes}
