\name{readBoxFile}
\alias{readBoxFile}
\alias{plot.TesseractBoxdata}
\title{Read a box file generated by tesseract}
\description{
  This is a simple function that reads the
  results of calling \code{tesseract ... makebox}.
  The result is a data frame, importantly, with a class
  \code{TesseractBoxdata}.
  This allows us to display the results via a method.
}
\usage{
readBoxFile(file, imgDim = integer(), ...)
\method{plot}{TesseractBoxdata}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{file}{the name of the file containing  the box data}
  \item{\dots}{additional arguments passed to \code{\link[utils]{read.table}}}
}
\value{
  This returns a data frame with the class \code{TesseractBoxdata}.
  The 5 columns are
  left, bottom, right, top of each box and \code{text}  contains the value.
  This allows us to easily plot the results.
}
\references{
Tesseract documentation.
}
\author{
Duncan Temple Lang
}


\examples{
readBoxFile(system.file("trainingSample", "eng.tables.exp0.box", package = "Rtesseract"))
plot(bb)
}
\keyword{IO}
\keyword{graphics}
