\name{tesseractVersion}
\alias{tesseractVersion}
\title{Get the version information of the tesseract library}
\description{
This returns the version information for the tesseract library
that is in use in this package.
}
\usage{
tesseractVersion()
}
\value{
 A character string (i.e., with a single element)
}
\references{
The Tesseract  project \url{https://code.google.com/p/tesseract-ocr/}
}
\author{
Duncan Temple Lang
}

%\seealso{}

\examples{
 tesseractVersion()
}
\keyword{programming}

